refactor: depart cases
