+++ packages/tests/src/e2e/bot/ProvisionCustomCopilotBasicBotForPythonUsingAZOAI.tests.ts
@@ -15,32 +15,16 @@ class AiBotOpenAITestCase extends CaseFactory {}
 
 // Azure OpenAI
 const myRecordAzOpenAI: Record<string, string> = {};
-myRecordAzOpenAI["custom-copilot-agent"] = "custom-copilot-agent-new";
 myRecordAzOpenAI["llm-service"] = "llm-service-azure-openai";
 myRecordAzOpenAI["azure-openai-key"] = "fake";
 myRecordAzOpenAI["azure-openai-deployment-name"] = "fake";
 myRecordAzOpenAI["azure-openai-endpoint"] = "https://test.com";
 new AiBotAzureOpenAITestCase(
-  Capability.Agent,
-  27689384,
+  Capability.AiBot,
+  27551399,
   "frankqian@microsoft.com",
   ["bot"],
   ProgrammingLanguage.PY,
   {},
   myRecordAzOpenAI
-).test();
-
-// OpenAI
-const myRecordOpenAI: Record<string, string> = {};
-myRecordOpenAI["custom-copilot-agent"] = "custom-copilot-agent-new";
-myRecordOpenAI["llm-service"] = "llm-service-openai";
-myRecordOpenAI["openai-key"] = "fake";
-new AiBotOpenAITestCase(
-  Capability.Agent,
-  27689385,
-  "frankqian@microsoft.com",
-  ["bot"],
-  ProgrammingLanguage.PY,
-  {},
-  myRecordOpenAI
 ).test();
