+++ packages/tests/src/e2e/bot/ProvisionCustomCopilotBasicBotForPythonUsingOAI.tests.ts
@@ -13,22 +13,6 @@ class AiBotAzureOpenAITestCase extends CaseFactory {}
 
 class AiBotOpenAITestCase extends CaseFactory {}
 
-// Azure OpenAI
-const myRecordAzOpenAI: Record<string, string> = {};
-myRecordAzOpenAI["llm-service"] = "llm-service-azure-openai";
-myRecordAzOpenAI["azure-openai-key"] = "fake";
-myRecordAzOpenAI["azure-openai-deployment-name"] = "fake";
-myRecordAzOpenAI["azure-openai-endpoint"] = "https://test.com";
-new AiBotAzureOpenAITestCase(
-  Capability.AiBot,
-  27551399,
-  "frankqian@microsoft.com",
-  ["bot"],
-  ProgrammingLanguage.PY,
-  {},
-  myRecordAzOpenAI
-).test();
-
 // OpenAI
 const myRecordOpenAI: Record<string, string> = {};
 myRecordOpenAI["llm-service"] = "llm-service-openai";
