+++ packages/tests/src/e2e/bot/ProvisionCustomCopilotRagBasicBotForPythonUsingAZOAI.tests.ts
@@ -29,18 +29,3 @@ new BasicRAGBotAzureOpenAITestCase(
   {},
   myRecordAzOpenAI
 ).test();
-
-// OpenAI
-const myRecordOpenAI: Record<string, string> = {};
-myRecordOpenAI["custom-copilot-rag"] = "custom-copilot-rag-customize";
-myRecordOpenAI["llm-service"] = "llm-service-openai";
-myRecordOpenAI["openai-key"] = "fake";
-new BasicRAGBotOpenAITestCase(
-  Capability.RAG,
-  27178104,
-  "frankqian@microsoft.com",
-  ["bot"],
-  ProgrammingLanguage.PY,
-  {},
-  myRecordOpenAI
-).test();
